+++ src/Dhall/Ex/Cmd/Run.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE DataKinds        #-}
+{-# LANGUAGE LambdaCase       #-}
 {-# LANGUAGE OverloadedLabels #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeOperators    #-}
@@ -9,6 +10,7 @@ module Dhall.Ex.Cmd.Run where
 import           RIO
 
 import           Data.Extensible
+import           Dhall.Ex.Config (Export, findExportByName)
 import           Dhall.Ex.Env
 
 showNotImpl :: MonadIO m => m ()
@@ -16,3 +18,13 @@ showNotImpl = hPutBuilder stdout "not yet implement command."
 
 class Run kv where
   run' :: proxy kv -> AssocValue kv -> RIO Env ()
+
+runWithOnly ::
+  (FilePath -> a -> Export -> RIO Env ()) -> FilePath -> a -> RIO Env ()
+runWithOnly act dir opts = do
+  config <- asks (view #config)
+  asks (view #only) >>= \case
+    Nothing   -> mapM_ (act dir opts) (config ^. #exports)
+    Just name -> case findExportByName (config ^. #exports) name of
+      Just export -> act dir opts export
+      Nothing     -> logError $ display ("undefined name: " <> name)