+++ src/Dhall/Ex/Config.hs
@@ -9,6 +9,7 @@
 module Dhall.Ex.Config where
 
 import           RIO
+import qualified RIO.List        as L
 
 import           Data.Extensible
 import qualified Dhall           as Dhall
@@ -31,3 +32,6 @@ readConfig path = (liftIO . Dhall.input Dhall.auto) =<< readFileUtf8 path
 
 ghToken :: FieldOptic "GH_TOKEN"
 ghToken = itemAssoc (Proxy @ "GH_TOKEN")
+
+findExportByName :: [Export] -> Text -> Maybe Export
+findExportByName exports name = L.find (\e -> e ^. #name == name) exports